add: use custom exceptions

+++ Lib.Application/UseCases/Authors/DeleteAuthorUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using Lib.Core.Abstractions;
+using Lib.Core.Exceptions;
 
 namespace Lib.Application.UseCases.Authors
 {
@@ -12,6 +13,14 @@ namespace Lib.Application.UseCases.Authors
 
         public async Task<Guid> ExecuteAsync(Guid id, CancellationToken cancellationToken)
         {
+
+            var author = await _unitOfWork.AuthorsRepository.GetAuthrorByIdAsync(id, cancellationToken);
+
+            if (author == null)
+            {
+                throw new NotFoundException($"Author with id {id} not found");
+            } 
+
             var authorId = await _unitOfWork.AuthorsRepository.DeleteAuthorAsync(id, cancellationToken);
 
             return authorId;
