+++ Lib.Application/UseCases/Books/DeleteBookUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using Lib.Core.Abstractions;
+using Lib.Core.Exceptions;
 
 namespace Lib.Application.UseCases.Books
 {
@@ -12,6 +13,14 @@ namespace Lib.Application.UseCases.Books
 
         public async Task<Guid> ExecuteAsync(Guid id, CancellationToken cancellationToken)
         {
+
+            var book = await _unitOfWork.BooksRepository.GetBookByIdAsync(id, cancellationToken);
+
+            if (book == null)
+            {
+                throw new NotFoundException($"Book with id {id} is not found");
+            }
+
             var guid = await _unitOfWork.BooksRepository.DeleteBookAsync(id, cancellationToken);
             return guid;
         }
